pre_open 파일 저장, pref에서 실행
