+++ release/scripts/startup/abler/general.py
@@ -195,10 +195,6 @@ class FileOpenOperator(bpy.types.Operator, ImportHelper):
         FILEPATH = self.filepath
         bpy.ops.wm.open_mainfile(filepath=FILEPATH)
 
-        # 아무 레이어나 변경했다가 다시 리셋
-        original_value = bpy.data.scenes["Scene"].ACON_prop.edge_min_line_width
-        bpy.data.scenes["Scene"].ACON_prop.edge_min_line_width = 0
-        bpy.data.scenes["Scene"].ACON_prop.edge_min_line_width = original_value
         return {"FINISHED"}
