--- release/scripts/startup/abler/lib/post_open.py
@@ -1,10 +1,0 @@
-import bpy
-from .tracker import tracker
-
-
-def tracker_file_open():
-
-    # tracking file_open
-    if bpy.data.filepath != "":
-        tracker.file_open()
-        return True
